+++ src/main/java/net/flameslight/magiccircles/datagen/types/transformations/TransformManager.java
@@ -4,7 +4,6 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import net.flameslight.magiccircles.datagen.types.EntitySnapshot;
 import net.flameslight.magiccircles.datagen.types.magicCircle.MagicCircleData;
 import net.flameslight.magiccircles.datagen.types.transformations.data.DataTransformExecutable;
-import net.flameslight.magiccircles.datagen.types.transformations.data.TempDataTransformExecutable;
 import net.flameslight.magiccircles.datagen.types.transformations.render.RenderTransformExecutable;
 
 import java.util.ArrayList;
@@ -16,10 +15,10 @@ import java.util.ArrayList;
  */
 public class TransformManager {
     // applied at circle init until a certain tick
-    private final ArrayList<TempDataTransformExecutable> initTransformations = new ArrayList<>();
+    private final ArrayList<DataTransformExecutable> initTransformations = new ArrayList<>();
 
     // applied at a certain time before circle termination until the circle termination
-    private final ArrayList<TempDataTransformExecutable> finalTransformations = new ArrayList<>();
+    private final ArrayList<DataTransformExecutable> finalTransformations = new ArrayList<>();
 
     // applied all the time until circle termination per tick
     private final ArrayList<DataTransformExecutable> dataTransformations = new ArrayList<>();
@@ -29,33 +28,35 @@ public class TransformManager {
 
     public void executeInitTransformations(EntitySnapshot entitySnapshot,
                                            MagicCircleData data,
-                                           float currentFullTicks,
-                                           int totalTicks) {
+                                           float ticksDifferenceFromLastCall,
+                                           float passedTransformFullTicks) {
         this.initTransformations.forEach(animation -> animation.execute(
                 entitySnapshot,
                 data,
-                currentFullTicks,
-                totalTicks));
+                ticksDifferenceFromLastCall,
+                passedTransformFullTicks));
     }
 
     public void executeFinalTransformations(EntitySnapshot entitySnapshot,
                                             MagicCircleData data,
-                                            float currentFullTicks,
-                                            int totalTicks) {
+                                            float ticksDifferenceFromLastCall,
+                                            float passedTransformFullTicks) {
         this.finalTransformations.forEach(animation -> animation.execute(
                 entitySnapshot,
                 data,
-                currentFullTicks,
-                totalTicks));
+                ticksDifferenceFromLastCall,
+                passedTransformFullTicks));
     }
 
     public void executePermanentDataTransformations(EntitySnapshot entitySnapshot,
                                                     MagicCircleData data,
-                                                    float currentFullTicks) {
+                                                    float ticksDifferenceFromLastCall,
+                                                    float passedTransformFullTicks) {
         this.dataTransformations.forEach(animation -> animation.execute(
                 entitySnapshot,
                 data,
-                currentFullTicks));
+                ticksDifferenceFromLastCall,
+                passedTransformFullTicks));
     }
 
     public void executeRenderTransformations(PoseStack poseStack,
@@ -69,11 +70,11 @@ public class TransformManager {
                 partialTicks));
     }
 
-    public void addInitTransformation(TempDataTransformExecutable executable) {
+    public void addInitTransformation(DataTransformExecutable executable) {
         this.initTransformations.add(executable);
     }
 
-    public void addFinalTransformation(TempDataTransformExecutable executable) {
+    public void addFinalTransformation(DataTransformExecutable executable) {
         this.finalTransformations.add(executable);
     }