+++ src/main/java/net/flameslight/magiccircles/datagen/types/transformations/data/DataTransformExecutable.java
@@ -7,5 +7,6 @@ import net.flameslight.magiccircles.datagen.types.magicCircle.MagicCircleData;
 public interface DataTransformExecutable {
     void execute(EntitySnapshot entitySnapshot,
                  MagicCircleData data,
-                 float newFullTicks);
+                 float ticksDifferenceFromLastCall,
+                 float passedTransformFullTicks);
 }